requirements.txt
Pyrogram
python-dotenv
TgCrypto